# docs/api/E2-identity/openapi.yaml
openapi: 3.1.0
info:
  title: hris-outsource · E2 Identity, Org & Master Data
  version: 1.0.0
  description: |
    OpenAPI 3.1 contract for **E2 — Identity, Org & Master Data**.

    > **NOTE — Profile change-requests removed 2026-06-14 (EPICS §8 E11):**
    > agent self-edits to phone, emergency contact, and bank account are now
    > **instant** (applied immediately, no approval). The `change-requests`
    > tag, paths, and schemas are deleted. Configurable approval now lives only
    > in E11 (leave/overtime).

    Scope: the people the system is about (employees + linked logins), the
    employment relationship (PKWT/PKWTT agreements + compensation), the client
    company directory, and the operational master data (leave types, attendance
    codes, overtime rules). Position is **free-text** (no master) with a
    typeahead search endpoint over existing values.

    Binding conventions: see `docs/api/CONVENTIONS.md` (sections referenced
    inline as §N). Picker endpoints are co-located with their owning resource
    per §18 — no separate `/lookups` namespace.

    Cross-references: employee-detail tab data (Penempatan/Kehadiran/Cuti/Lembur)
    is **not served** by E2; the E2 employee-detail response carries
    `x-cross-epic-references` mapping each tab to the owning epic's endpoint.

    Personas used in examples: **Sari Hadi (HR Admin)**, **Rudi Wijaya (Shift
    Leader, Plaza Senayan)**, **Budi Santoso (Agent, Parking Attendant)**.
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /positions:search:
    get:
      operationId: searchPositions
      tags: [people]
      summary: Position typeahead — distinct existing position strings
      description: |
        Position is **free-text** (no master entity, no FK, no ID). This
        endpoint backs the position typeahead on the placement create/transfer
        forms (E3): given a query `q`, it returns the **distinct** existing
        position strings (case-insensitive substring match) already recorded
        across placements, so admins can reuse a consistent label or type a new
        one. There is no create/update/delete — values are persisted as the
        free-text `position` on the placement (E3).
      x-rbac:
        roles: [hr_admin, super_admin, shift_leader]
        scope: company_or_global
      parameters:
        - in: query
          name: q
          description: |
            Substring to match against existing position strings
            (case-insensitive). Omit to get the most common positions.
          schema: { type: string }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Distinct matching position strings.
          content:
            application/json:
              schema:
                type: object
                required: [positions]
                properties:
                  positions:
                    type: array
                    items: { type: string }
              examples:
                parking_match:
                  value:
                    positions:
                      - Petugas Parkir
                      - Koordinator Lokasi
                      - Supervisor Parkir
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # =====================================================================
  # OPERATIONAL MASTER DATA (F2.5)
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-jwt
      description: |
        `Authorization: Bearer <token>` per CONVENTIONS §3. Token type
        (opaque vs JWT) is deferred to E1; consumers MUST treat as opaque.
  parameters:
    Limit:
      in: query
      name: limit
      description: Page size (default 50, max 200) per §8.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: Malformed request or invalid input (CONVENTIONS §11).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            generic:
              value:
                error:
                  code: INVALID_REQUEST
                  message: "Permintaan tidak valid."
                  request_id: req_01J5XK0A
    Unauthenticated:
      description: Missing or expired bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            expired:
              value:
                error:
                  code: UNAUTHENTICATED
                  message: "Sesi berakhir. Silakan masuk kembali."
                  request_id: req_01J5XK0B
    Forbidden:
      description: Authenticated but role-not-permitted (CONVENTIONS §3, §17).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            no_permission:
              value:
                error:
                  code: FORBIDDEN
                  message: "Anda tidak memiliki izin untuk tindakan ini."
                  request_id: req_01J5XK0C
...
  schemas:
...
    ErrorEnvelope:
      type: object
      description: Standard error envelope (CONVENTIONS §11).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: "QUOTA_EXCEEDED" }
            message: { type: string, example: "Permintaan melebihi kuota tersisa." }
            fields:
              type: object
              additionalProperties: { type: string }
              description: Only on 400/422 — maps field name → field-level error.
            request_id: { type: string, example: "req_01J5XK..." }
...